// src/interfaces/spotify/spotify-artist.interface.ts
export interface MerchItem {
  id: string;
  uri: string;
  name: string;
  type: LatestType;
  label: any;
  copyright: Copyright;
  date: PurpleDate;
  coverArt: CoverArt;
  tracks: Tracks;
  playability: Playability;
  sharingInfo: SharingInfo;
}

export interface Copyright {
  items: CopyrightItem[];
}

export interface CopyrightItem {
  type: PurpleType;
  text: string;
}

export enum PurpleType {
  C = 'C',
  P = 'P',
}

export interface CoverArt {
  sources: ItemSource[];
}

export interface ItemSource {
  url: string;
  width: number | null;
  height: number | null;
}

export interface PurpleDate {
  year: number;
  month: number;
  day: number;
  precision: Precision;
}

export enum Precision {
  Day = 'DAY',
}

export interface Playability {
  playable: boolean;
  reason: Reason;
}

export enum Reason {
  Playable = 'PLAYABLE',
}

export interface SharingInfo {
  shareId: string;
  shareUrl: string;
}

export interface Tracks {
  totalCount: number;
}

export enum LatestType {
  Album = 'ALBUM',
  Compilation = 'COMPILATION',
  Ep = 'EP',
  Single = 'SINGLE',
}

export interface Latest {
  id: string;
  uri: string;
  name: string;
  type: LatestType;
  copyright: Copyright;
  date: LatestDate;
  coverArt: CoverArt;
  tracks: Tracks;
  label: any;
  playability: Playability;
}

export interface LatestDate {
  year: number;
}

export interface TopTracks {
  items: TopTracksItem[];
}

export interface TopTracksItem {
  uid: string;
  track: Track;
}

export interface Track {
  id: string;
  uri: string;
  name: string;
  playcount: string;
  discNumber: number;
  duration: Duration;
  playability: Playability;
  contentRating: ContentRating;
  artists: TrackArtists;
  album: Album;
}

export interface Album {
  uri: string;
  coverArt: AlbumCoverArt;
}

export interface AlbumCoverArt {
  sources: PurpleSource[];
}